SFR-2277: Renaming process files and refiling ingest processes
