+++ processes/__init__.py
@@ -1,22 +1,22 @@
 from .covers import CoverProcess
-from .doab import DOABProcess
-from .gutenberg import GutenbergProcess
-from .hathiTrust import HathiTrustProcess
-from .nypl import NYPLProcess
-from .oclcCatalog import CatalogProcess
-from .oclcClassify import ClassifyProcess
+from .ingest.doab import DOABProcess
+from .ingest.gutenberg import GutenbergProcess
+from .ingest.hathi_trust import HathiTrustProcess
+from .ingest.nypl import NYPLProcess
+from .catalog import CatalogProcess
+from .classify import ClassifyProcess
 from .sfrCluster import ClusterProcess
 from .local_development_setup import LocalDevelopmentSetupProcess
 from .seed_local_data import SeedLocalDataProcess
-from .s3Files import S3Process
+from .s3_files import S3Process
 from .api import APIProcess
-from .muse import MUSEProcess
+from .ingest.muse import MUSEProcess
 from .ingestReport import IngestReportProcess
-from .met import METProcess
+from .ingest.met import METProcess
 from .db_maintenance import DatabaseMaintenanceProcess
 from .db_migration import MigrationProcess
-from .chicagoISAC import ChicagoISACProcess
-from .UofSC import UofSCProcess
-from .loc import LOCProcess
-from .UofM import UofMProcess
+from .ingest.chicago_isac import ChicagoISACProcess
+from .ingest.u_of_sc import UofSCProcess
+from .ingest.loc import LOCProcess
+from .ingest.u_of_m import UofMProcess
 from .fulfill_url_manifest import FulfillURLManifestProcess
