+++ processes/ingest/muse.py
@@ -6,7 +6,7 @@ from pymarc import MARCReader
 import requests
 from requests.exceptions import ReadTimeout, HTTPError
 
-from .core import CoreProcess
+from ..core import CoreProcess
 from mappings.muse import MUSEMapping
 from managers import MUSEError, MUSEManager
 from logger import createLog
